chore: rename camelcased files to kebabcase
